requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alpaca_eval==0.6.2
altair==5.3.0
annotated-types==0.6.0
anthropic==0.19.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.0.5
attrs==23.2.0
blobfile==2.1.1
boto3==1.34.124
botocore==1.34.124
cachetools==5.3.3
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.25.0
cohere==5.6.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cycler==0.12.1
DataProperty==1.0.1
datasets==3.2.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
docstring_parser==0.16
email_validator==2.1.1
evaluate==0.4.2
executing==0.8.3
fastapi==0.111.0
fastapi-cli==0.0.4
fastavro==1.9.4
ffmpy==0.3.2
filelock==3.14.0
fire==0.6.0
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.1
gradio==4.32.2
gradio_client==0.17.0
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.7
importlib_resources==6.4.0
interegular==0.3.3
IOUtils==1.0.3
ipykernel==6.28.0
ipython==8.20.0
jedi==0.18.1
Jinja2==3.1.4
jiter==0.4.1
jmespath==1.0.1
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lark==1.1.9
lion-pytorch==0.1.2
lit==15.0.7
llvmlite==0.42.0
lm-format-enforcer==0.10.3
lxml==4.9.4
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mbstrdecoder==1.1.3
mdurl==0.1.2
mistralai==0.4.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.17
ninja==1.11.1.1
nltk==3.8.1
numba==0.59.1
numexpr==2.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.30.5
orjson==3.10.3
outlines==0.0.46
packaging==24.0
pandas==2.2.2
parameterized==0.9.0
parso==0.8.3
pathvalidate==3.2.0
patsy==0.5.6
peft==0.8.2
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.2.2
portalocker==2.8.2
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycountry==24.6.1
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic-settings==2.3.1
pydantic_core==2.18.2
pydub==0.25.1
pygments==2.15.1
pyparsing==3.1.2
pytablewriter==1.2.0
python-dateutil==2.9.0post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
ray==2.23.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
responses==0.18.0
rich==13.7.1
rouge_score==0.1.2
rpds-py==0.18.1
rsa==4.9
ruff==0.4.7
s3transfer==0.10.1
sacrebleu==2.4.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.1.1
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlitedict==2.1.0
stack-data==0.2.0
starlette==0.37.2
svgwrite==1.4.3
sympy==1.12.1
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomlkit==0.12.0
toolz==0.12.1
torchao==0.3.1
torchtune @ git+https://github.com/pytorch/torchtune.git@f0a15c5597340efc7ad9e67b4770b6693547d143
tornado==6.3.3
tqdm==4.66.4
tqdm-multiprocess==0.0.11
traitlets==5.7.1
transformers==4.43.3
triton==2.3.1
trl==0.9.6
typepy==1.3.2
typer==0.12.3
types-requests==2.32.0.20240602
typing_extensions==4.12.1
tyro==0.8.5
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
vllm==0.5.3.post1
vllm-flash-attn==2.5.9.post1
wandb==0.16.0
watchfiles==0.22.0
wavedrom==2.0.3.post3
wcwidth==0.2.5
websockets==11.0.3
word2number==1.1
xformers==0.0.27
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0
